requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.8.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0

# Elasticsearch and search dependencies
elasticsearch==8.15.1
elasticsearch-dsl==8.16.0

# Google Cloud and Vertex AI
google-cloud-aiplatform==1.75.0
google-auth==2.39.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Audio processing
librosa==0.10.2.post1
soundfile==0.12.1
torch==2.5.1
torchaudio==2.5.1

# Text embeddings and transformers
transformers==4.46.3
sentence-transformers==3.3.1

# Environment and configuration
python-dotenv==1.0.1
PyYAML==6.0.2

# Audio recorder for Streamlit (if needed)
audio-recorder-streamlit==0.0.10
